Move query to users repo
