+++ internal/handlers/posts/handlers.go
@@ -389,7 +389,7 @@ func (s *Service) SinglePostHandler(w http.ResponseWriter, r *http.Request) {
 
 	// Check whether the current user liked and/or faved the post
 	if data.CurrentUser.IsAuthenticated() {
-		userActions, _ := s.postsRepo.GetUserActions(
+		userActions, _ := s.usersRepo.GetUserActions(
 			r.Context(),
 			data.CurrentUser.ID,
 			data.CurrentPost.ID,
