+++ internal/handlers/posts/posts.go
@@ -6,11 +6,13 @@ import (
 	"github.com/vlatan/video-store/internal/integrations/gemini"
 	"github.com/vlatan/video-store/internal/integrations/yt"
 	postsRepo "github.com/vlatan/video-store/internal/repositories/posts"
+	usersRepo "github.com/vlatan/video-store/internal/repositories/users"
 	"github.com/vlatan/video-store/internal/ui"
 )
 
 type Service struct {
 	postsRepo *postsRepo.Repository
+	usersRepo *usersRepo.Repository
 	rdb       *rdb.Service
 	ui        ui.Service
 	config    *config.Config
@@ -20,6 +22,7 @@ type Service struct {
 
 func New(
 	postsRepo *postsRepo.Repository,
+	usersRepo *usersRepo.Repository,
 	rdb *rdb.Service,
 	ui ui.Service,
 	config *config.Config,
@@ -28,6 +31,7 @@ func New(
 ) *Service {
 	return &Service{
 		postsRepo: postsRepo,
+		usersRepo: usersRepo,
 		rdb:       rdb,
 		ui:        ui,
 		config:    config,
