+++ internal/repositories/users/queries.go
@@ -40,3 +40,15 @@ const getUsersQuery = `
 	ORDER BY created_at
 	LIMIT $1 OFFSET $2
 `
+
+const userActionsQuery = `
+	SELECT 
+		EXISTS (
+			SELECT 1 FROM post_like
+			WHERE user_id = $1 AND post_id = $2
+		) AS liked,
+		EXISTS (
+			SELECT 1 FROM post_fave
+			WHERE user_id = $1 AND post_id = $2
+		) AS faved
+`
